fix create_gmap return code
Function create_gmap() returns the fd on success or -1 on error,
but its caller was assuming !0 on success or 0 on error.  There was
also a distinct lack of error condition passing.

Introduce a new function: dt_bpf_error().

Fix create_gmap() to set errno correctly, along with a descriptive
error message.

Fix dt_bpf_gmap_create() to correctly check the return value from
calls to create_gmap().

Fix dt_bpf_gmap_create() to not call create_gmap() for maps with
zero elements.  Trying to create a BPF map with zero elements results
in an error.  If the requested number of elements is zero, the BPF
map need not be created because it will not be used.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>