Interrupting tracing should not yield an error (unless we failed)
Interrupting tracing (e.g. using Ctrl-C) would yield an error that
tracing couldn't be stopped.  This was caused by a remnant of the
old dtrace code using an ioctl() to signal the kernel that tracing
should be stopped.

This error condition no longer occurs because we no longer use the
ioxtl().  Additional processing when tracing should be stopped is
yet to be implemented.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>