Enforce error and drop tag verification for tests.
Various tests exercise error and probe drop conditions and request
that a representative error or drop tag is included in the output.
Tests that are named err.* will be executed with the -xerrtags option
and drp.* tests add the -xdroptags option.

Up until now there was no explicit verification in runtest.sh to
ensure that a test reports the correct tag.  This patch adds that
verification.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>