Remove obsolete dt_fd, dt_ftfd, and dt_fterr
Given that DTrace no longer makes use of the ioctl() interface to a
DTrace kernel component  we can remove all references to these members
in dtrace_hdl_t.  We retain dtrace_ctlfd() because it is part of the
public API, and dt_ioctl() because various parts of libdtrace still
need to rewritten to replace the ioctl() use.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>