Change the dtrace_eprobedesc_t structure to match the new design
In DTrace v1, the dtrace_eprobedesc_t structure was used to obtain
metadata information about the probe data stream from the kernel.
Each EPID was also associated with a speciifc dtrace_probedesc_t,
identifying the probe for which the EPID represents an enabling.
Each EPID identified a unique dtrace_eprobedesc_t while the
dtrace_probedesc_t could be shared between multiple EPIDs.

    EPID -[1-to-1]-> dtrace_eprobedesc_t
    EPID -[n-to-1]-> dtrace_probedesc_t

In the DTrace v2 design, where everything resides in userspace, we
can create that metadata directly.  We can also make use of the fact
that when a single program is attached to multiple probes, we do not
need more than one copy of the metadata.  We also make use of the fact
that the EPID identifies a unique (probe, program) tuple.

    EPID -[n-to-1]-> dtrace_datadesc_t
    EPID -[n-to-1]-> dtrace_probedesc_t

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>